+++ addons/bus/__manifest__.py
@@ -33,7 +33,7 @@
         ],
         'bus.websocket_worker_assets': [
             'web/static/src/legacy/js/promise_extension.js',
-            'web/static/src/boot.js',
+            'web/static/src/module_loader.js',
             'bus/static/src/workers/*',
         ],
     },
