+++ addons/mail/views/discuss_public_templates.xml
@@ -12,6 +12,7 @@
                         __session_info__: <t t-out="json.dumps(session_info)"/>,
                         csrf_token: "<t t-out="request.csrf_token(None)"/>",
                         debug: "<t t-out="debug"/>",
+                        discuss_data: <t t-out="json.dumps(data)"/>
                     };
                     {
                         const { user_context,  cache_hashes } = odoo.__session_info__;
@@ -25,13 +26,6 @@
                         direction: <t t-out="env['res.lang']._lang_get(lang or env.user.lang).direction or 'ltr'" />;
                     }
                 </style>
-                <script>
-                    odoo.define('mail.discuss_public_template', [], function() {
-                        return {
-                            data: <t t-out="json.dumps(data)"/>,
-                        };
-                    });
-                </script>
             </head>
             <body class="d-flex flex-column h-100">
                 <t t-out="body or ''"/>
