+++ addons/project/__manifest__.py
@@ -85,7 +85,6 @@
         ],
         'web.assets_tests': [
             'project/static/tests/tours/**/*',
-            'project/static/tests/views/**/*',
         ],
         'project.webclient': [
             ('include', 'web._assets_helpers'),
@@ -111,7 +110,7 @@
             'web/static/src/scss/fontawesome_overridden.scss',
 
             'web/static/src/legacy/js/promise_extension.js',
-            'web/static/src/boot.js',
+            'web/static/src/module_loader.js',
             'web/static/src/session.js',
             'web/static/src/legacy/js/core/cookie_utils.js',
