+++ addons/survey/static/tests/tours/certification_failure.js
@@ -1,15 +1,17 @@
 /** @odoo-module **/
 
-import SurveyFormWidget from "@survey/js/survey_form";
 /**
  * Speed up fade-in fade-out to avoid useless delay in tests.
- */
-SurveyFormWidget.include({
-    _submitForm: function () {
-        this.fadeInOutDelay = 0;
-        return this._super.apply(this, arguments);
-    }
-});
+*/
+function patchSurveyWidget() {
+    const SurveyFormWidget = odoo.loader.modules.get('@survey/js/survey_form')[Symbol.for('default')]
+    SurveyFormWidget.include({
+        _submitForm: function () {
+            this.fadeInOutDelay = 0;
+            return this._super.apply(this, arguments);
+        }
+    });
+}
 
 /**
  * This tour will test that, for the demo certification allowing 2 attempts, a user can
@@ -17,6 +19,14 @@ SurveyFormWidget.include({
  */
 
 import { registry } from "@web/core/registry";
+
+var patch = [{
+    content: "Patching Survey Widget",
+    trigger: 'body',
+    run: function(){
+        patchSurveyWidget();
+    }
+}]
 
 var failSteps = [{ // Page-1
     content: "Clicking on Start Certification",
@@ -104,4 +114,4 @@ var lastSteps = [{
 registry.category("web_tour.tours").add('test_certification_failure', {
     test: true,
     url: '/survey/start/4ead4bc8-b8f2-4760-a682-1fde8daaaaac',
-    steps: () => [].concat(failSteps, retrySteps, failSteps, lastSteps) });
+    steps: () => [].concat(patch, failSteps, retrySteps, failSteps, lastSteps) });
