+++ addons/test_website/__manifest__.py
@@ -28,7 +28,7 @@ models which only purpose is to run tests.""",
             'http://test.external.link/javascript1.js',
             '/web/static/lib/jquery.ui/jquery-ui.css',
             'http://test.external.link/style1.css',
-            '/web/static/src/boot.js',
+            '/web/static/src/module_loader.js',
             'http://test.external.link/javascript2.js',
             'http://test.external.link/style2.css',
         ],
