+++ addons/web/static/src/core/emoji_picker/emoji_picker.js
@@ -115,7 +115,7 @@ export const loader = {
 export async function loadEmoji() {
     try {
         await loader.loadEmoji();
-        return odoo.runtimeImport("@web/core/emoji_picker/emoji_data");
+        return odoo.loader.modules.get("@web/core/emoji_picker/emoji_data");
     } catch {
         // Could be intentional (tour ended successfully while emoji still loading)
         return { emojis: [], categories: [] };
