--- addons/website/static/src/client_actions/website_preview/website_preview_test_mode.js
@@ -1,12 +1,0 @@
-/** @odoo-module **/
-import { patch } from "@web/core/utils/patch";
-import { WebsitePreview } from '@website/client_actions/website_preview/website_preview';
-
-patch(WebsitePreview.prototype, {
-    /**
-     * @override
-     */
-    get testMode() {
-        return true;
-    }
-});
