+++ addons/website/static/src/js/content/website_root_instance.js
@@ -2,15 +2,15 @@
 
 import { createPublicRoot } from "@web/legacy/js/public/public_root";
 import { WebsiteRoot } from "./website_root";
-import { loadLegacyWysiwygAssets } from "@web_editor/js/frontend/loader";
+import { getBundle, loadBundle } from "@web/core/assets";
 
-export default createPublicRoot(WebsiteRoot).then(rootInstance => {
+export default createPublicRoot(WebsiteRoot).then(async (rootInstance) => {
     // This data attribute is set by the WebsitePreview client action for a
     // restricted editor user.
     if (window.frameElement && window.frameElement.dataset.loadWysiwyg === 'true') {
-        loadLegacyWysiwygAssets(['web_editor.assets_wysiwyg', 'website.assets_wysiwyg']).then(() => {
-            window.dispatchEvent(new CustomEvent('PUBLIC-ROOT-READY', {detail: {rootInstance}}));
-        });
+        const assets = await getBundle("website.assets_all_wysiwyg");
+        await loadBundle(assets);
+        window.dispatchEvent(new CustomEvent('PUBLIC-ROOT-READY', {detail: {rootInstance}}));
     }
     return rootInstance;
 });
