+++ addons/website/static/tests/tours/public_user_editor_dep_widget.js
@@ -1,16 +1,18 @@
-/** @odoo-module **/
+odoo.loader.bus.addEventListener("module-started", (e) => {
+    if (e.detail.moduleName === "@web_editor/js/frontend/loadWysiwygFromTextarea") {
+        const publicWidget = odoo.loader.modules.get("@web/legacy/js/public/public_widget")[Symbol.for('default')];
+        const { loadWysiwygFromTextarea } = e.detail.module;
 
-import publicWidget from "@web/legacy/js/public/public_widget";
-import { loadWysiwygFromTextarea } from "@web_editor/js/frontend/loadWysiwygFromTextarea";
+        publicWidget.registry['public_user_editor_test'] = publicWidget.Widget.extend({
+            selector: 'textarea.o_public_user_editor_test_textarea',
 
-publicWidget.registry['public_user_editor_test'] = publicWidget.Widget.extend({
-    selector: 'textarea.o_public_user_editor_test_textarea',
-
-    /**
-     * @override
-     */
-    start: async function () {
-        await this._super(...arguments);
-        await loadWysiwygFromTextarea(this, this.el, {});
-    },
-});
+            /**
+             * @override
+             */
+            start: async function () {
+                await this._super(...arguments);
+                await loadWysiwygFromTextarea(this, this.el, {});
+            },
+        });
+    }
+})
