+++ addons/website_blog/__manifest__.py
@@ -32,10 +32,10 @@
     'assets': {
         'website.assets_wysiwyg': [
             'website_blog/static/src/js/options.js',
-            'website_blog/static/src/js/wysiwyg.js',
             'website_blog/static/src/snippets/s_blog_posts/options.js',
         ],
         'website.assets_editor': [
+            'website_blog/static/src/js/wysiwyg_adapter.js',
             'website_blog/static/src/js/tours/website_blog.js',
             'website_blog/static/src/js/systray_items/*.js',
         ],
