+++ addons/website_livechat/static/tests/tours/website_livechat_request.js
@@ -3,30 +3,34 @@
 import { patch } from "@web/core/utils/patch";
 import { endDiscussion, okRating, feedback, transcript, close } from "./website_livechat_common";
 import { registry } from "@web/core/registry";
-import { LivechatService, ODOO_VERSION_KEY } from "@im_livechat/embed/core/livechat_service";
 
-// Livechat sessions are deleted if they come from a different odoo version than
-// the current one. Since each test is run with a different browser the version
-// is never set. This patch sets the version to the current one.
-patch(LivechatService.prototype, {
-    async initialize() {
-        try {
-            const response = await fetch("/web/webclient/version_info", {
-                method: "POST",
-                headers: {
-                    "Content-Type": "application/json",
-                },
-                body: JSON.stringify({}),
-            });
-            const { server_version: serverVersion } = (await response.json()).result;
-            localStorage.setItem(ODOO_VERSION_KEY, serverVersion);
-        } catch {
-            // Could be intentional (tour ended successfully while version still loading)
-        }
-
-        return super.initialize(...arguments);
-    },
-});
+odoo.loader.bus.addEventListener("module-started", (e) => {
+    if (e.detail.moduleName === "@im_livechat/embed/core/livechat_service") {
+        const { LivechatService, ODOO_VERSION_KEY } = e.detail.module;
+        // Livechat sessions are deleted if they come from a different odoo version than
+        // the current one. Since each test is run with a different browser the version
+        // is never set. This patch sets the version to the current one.
+        patch(LivechatService.prototype, {
+            async initialize() {
+                try {
+                    const response = await fetch("/web/webclient/version_info", {
+                        method: "POST",
+                        headers: {
+                            "Content-Type": "application/json",
+                        },
+                        body: JSON.stringify({}),
+                    });
+                    const { server_version: serverVersion } = (await response.json()).result;
+                    localStorage.setItem(ODOO_VERSION_KEY, serverVersion);
+                } catch {
+                    // Could be intentional (tour ended successfully while version still loading)
+                }
+    
+                return super.initialize(...arguments);
+            },
+        });
+    }
+})
 
 const chatRequest = [
     {
