+++ addons/website_sale/static/tests/tours/website_sale_google_analytics.js
@@ -1,20 +1,25 @@
 /** @odoo-module **/
 
 import { registry } from "@web/core/registry";
-import websiteSaleTracking from "@website_sale/js/website_sale_tracking";
 import tourUtils from "@website_sale/js/tours/tour_utils";
+
+odoo.loader.bus.addEventListener("module-started", (e) => {
+    if (e.detail.moduleName === "@website_sale/js/website_sale_tracking") {
+        //import websiteSaleTracking from "@website_sale/js/website_sale_tracking";
+        e.detail.module[Symbol.for("default")].include({
+            // Purposely don't call super to avoid call to third party (GA) during tests
+            _onViewItem(event, data) {
+                $('body').attr('view-event-id', data.item_id);
+            },
+            _onAddToCart(event, data) {
+                $('body').attr('cart-event-id', data.item_id);
+            },
+        });
+    }
+});
 
 let itemId;
 
-websiteSaleTracking.include({
-    // Purposely don't call super to avoid call to third party (GA) during tests
-    _onViewItem(event, data) {
-        $('body').attr('view-event-id', data.item_id);
-    },
-    _onAddToCart(event, data) {
-        $('body').attr('cart-event-id', data.item_id);
-    },
-});
 
 registry.category("web_tour.tours").add('google_analytics_view_item', {
     test: true,
