+++ addons/website_slides/__manifest__.py
@@ -149,7 +149,7 @@ Featuring
             'website/static/src/libs/zoomodoo/zoomodoo.scss',
 
             'web/static/src/legacy/js/promise_extension.js',
-            'web/static/src/boot.js',
+            'web/static/src/module_loader.js',
             'web/static/src/session.js',
             'web/static/src/legacy/js/core/cookie_utils.js',
